refactor(shared): extract prompt helpers into shared.prompt_utils
Prompt assembly and tool-server prompt injection are common agent-loop behavior. Centralize them so plugins build system prompts consistently.
