+++ src/owui_ext/tools/multi_model_council.py
@@ -18,6 +18,11 @@ from pydantic import BaseModel, Field
 
 from owui_ext.shared.async_utils import maybe_await
 from owui_ext.shared.builtin_tools import BUILTIN_TOOL_CATEGORIES, VALVE_TO_CATEGORY
+from owui_ext.shared.prompt_utils import (
+    _append_tool_server_prompts,
+    merge_prompt_sections,
+    truncate_text,
+)
 from owui_ext.shared.terminal_events import emit_terminal_tool_event
 from owui_ext.shared.tool_event_metadata import CITATION_TOOLS, TERMINAL_EVENT_TOOLS
 
@@ -227,14 +232,6 @@ def build_direct_tools_dict(*, tool_servers: List[dict]) -> Dict[str, dict]:
     return direct_tools
 
 
-def truncate_text(value: str, limit: int = 200) -> str:
-    if not value:
-        return ""
-    if len(value) <= limit:
-        return value
-    return value[: limit - 3] + "..."
-
-
 def coerce_user_valves(raw_valves: Any, valves_cls: Type[BaseModel]) -> BaseModel:
     if isinstance(raw_valves, valves_cls):
         return raw_valves
@@ -409,18 +406,6 @@ def build_council_system_prompt(base_prompt: str, include_sources: bool) -> str:
     if base_prompt:
         return f"{base_prompt}\n\n{council_prompt}"
     return council_prompt
-
-
-def merge_prompt_sections(*sections: Optional[str]) -> str:
-    """Join non-empty prompt sections with blank lines."""
-    merged_sections = []
-    for section in sections:
-        if not isinstance(section, str):
-            continue
-        stripped = section.strip()
-        if stripped:
-            merged_sections.append(stripped)
-    return "\n\n".join(merged_sections)
 
 
 def normalize_terminal_tools_result(*, terminal_tools_result: Any, extra_params: Optional[dict]) -> dict:
@@ -693,50 +678,6 @@ async def apply_inlet_filters_if_enabled(
     except Exception as exc:
         log.warning(f"Error applying inlet filters: {exc}")
 
-    return form_data
-
-
-def _append_tool_server_prompts(form_data: dict, extra_params: dict) -> dict:
-    """Append terminal/direct-tool-server system prompts to messages.
-
-    Open WebUI core injects these prompts AFTER inlet filters so they survive
-    filters that rewrite the system message.  We replicate the same ordering by
-    calling this helper after ``apply_inlet_filters_if_enabled``.
-    """
-    prompts: list[str] = []
-    terminal_prompt = (extra_params or {}).get("__terminal_system_prompt__")
-    if isinstance(terminal_prompt, str) and terminal_prompt.strip():
-        prompts.append(terminal_prompt)
-    direct_prompts = (extra_params or {}).get(
-        "__direct_tool_server_system_prompts__", []
-    )
-    if isinstance(direct_prompts, list):
-        prompts.extend(p for p in direct_prompts if isinstance(p, str) and p.strip())
-
-    if not prompts:
-        return form_data
-
-    messages = list(form_data.get("messages", []))
-    combined = "\n\n".join(prompts)
-    if messages and messages[0].get("role") == "system":
-        # Copy the dict so we don't mutate current_messages across iterations
-        msg = {**messages[0]}
-        content = msg.get("content", "")
-        # Handle structured content (list of {"type": "text", "text": ...})
-        # the same way core's update_message_content does.
-        if isinstance(content, list):
-            msg["content"] = [
-                {**item, "text": f'{item["text"]}\n{combined}'}
-                if item.get("type") == "text"
-                else item
-                for item in content
-            ]
-        else:
-            msg["content"] = f"{content}\n\n{combined}" if content else combined
-        messages[0] = msg
-    else:
-        messages.insert(0, {"role": "system", "content": combined})
-    form_data["messages"] = messages
     return form_data
